# docs/swagger.yaml
basePath: /
definitions:
  dto.AcidCreateRequest:
    properties:
      hplus:
        minimum: 1
        type: integer
      info:
        maxLength: 500
        type: string
      molar_mass:
        minimum: 0
        type: number
      name:
        maxLength: 15
        type: string
      name_ext:
        maxLength: 25
        type: string
    required:
    - hplus
    - molar_mass
    - name
    - name_ext
    type: object
  dto.AcidListResponse:
    properties:
      acids:
        items:
          $ref: '#/definitions/dto.AcidResponse'
        type: array
    type: object
  dto.AcidResponse:
    properties:
      hplus:
        type: integer
      id:
        type: integer
      img:
        type: string
      info:
        type: string
      molar_mass:
        type: number
      name:
        type: string
      name_ext:
        type: string
    type: object
  dto.AcidUpdateRequest:
    properties:
      hplus:
        minimum: 1
        type: integer
      info:
        maxLength: 500
        type: string
      molar_mass:
        minimum: 0
        type: number
      name:
        maxLength: 15
        type: string
      name_ext:
        maxLength: 25
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: kovalyovea@student.bmstu.ru
    name: API Support
    url: https://github.com/nylox23/DIA
  description: Calculates the volume of gas emitted in a reaction
  license:
    name: AS IS (NO WARRANTY)
  title: Carb_Calc
  version: "1.0"
paths:
  /api/acids:
    get:
      description: Возвращает список кислот, отфильтрованный по названию
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AcidListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список кислот
      tags:
      - acids
    post:
      consumes:
      - application/json
      description: Создает новую кислоту без изображения
      security:
        - Bearer: []
      parameters:
      - description: Данные создавемой кислоты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcidCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AcidResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать кислоту
      tags:
      - acids
  /api/acids/{id}:
    delete:
      description: Проводит "удаление" кислоты, устанавливая флаг is_delete в true
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить кислоту
      tags:
      - acids
    get:
      description: Возвращает данные кислоты по ID
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AcidResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить кислоту по ID
      tags:
      - acids
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей кислоты
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые данные кислоты
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AcidUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AcidResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить кислоту
      tags:
      - acids
  /api/acids/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для кислоты, замещая предыдущее, если есть
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      - description: Файл с изображением
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить изображение кислоты
      tags:
      - acids
  /api/acids/{id}/toCarbonate:
    post:
      description: Добавляет кислоту в текущую заявку-черновик пользователя
      security:
        - Bearer: []
      parameters:
      - description: ID кислоты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить кислоту в заявку
      tags:
      - acids
...
schemes:
- http
swagger: "2.0"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header